// app/types/strapi/base/strapi-image.ts
export interface FormatAttributes {
  ext: string;
  url: string;
  hash: string;
  mime: string;
  name: string;
  path?: any;
  size: number;
  width: number;
  height: number;
}

export interface Thumbnail {
  ext: string;
  url: string;
  hash: string;
  mime: string;
  name: string;
  path?: any;
  size: number;
  width: number;
  height: number;
}

export interface Formats {
  large: FormatAttributes;
  small: FormatAttributes;
  medium: FormatAttributes;
  thumbnail: Thumbnail;
}

export interface BgImage {
  name: string;
  alternativeText: string;
  caption: string;
  width: number;
  height: number;
  formats: Formats;
  hash: string;
  ext: string;
  mime: string;
  size: number;
  url: string;
  previewUrl?: any;
  provider: string;
  provider_metadata?: any;
  createdAt: Date;
  updatedAt: Date;
}

export interface ImageAttributes {
  id: number;
  attributes: BgImage;
}

export interface BgImageObject {
  data: ImageAttributes[];
}